// src/web/src/api/generated/models/ActivitySessionSignalsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ActivitySessionSignalsResponse = {
    pain: number | string | null;
    fatigue: number | string | null;
    sleepQuality: number | string | null;
    perceivedRecovery: number | string | null;
    gaitChanged: boolean | null;
    hasIllnessOrSymptom: boolean | null;
    recoveryResponse: string | null;
    painLocation: string | null;
};